requirements-ref.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
alembic==1.14.0
altair==5.4.1
annotated-types==0.7.0
anyio==4.6.2.post1
appnope==0.1.4
asttokens==2.4.1
attrs==24.2.0
awscli==1.35.8
beautifulsoup4==4.12.3
blinker==1.9.0
botocore==1.35.42
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cloudpickle==3.1.0
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
databricks-sdk==0.38.0
debugpy==1.8.7
decorator==5.1.1
Deprecated==1.2.15
deprecation==2.1.0
docker==7.1.0
docutils==0.16
executing==2.1.0
fastapi==0.115.2
Flask==3.1.0
fonttools==4.55.0
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.36.0
gotrue==2.9.3
graphene==3.4.3
graphql-core==3.2.5
graphql-relay==3.2.0
gunicorn==23.0.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.6
httpx==0.27.2
hyperframe==6.0.1
idna==3.10
importlib_metadata==8.5.0
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.28.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
lxml==5.3.0
Mako==1.3.6
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.1
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mlflow==2.18.0
mlflow-skinny==2.18.0
multidict==6.1.0
narwhals==1.9.3
nest-asyncio==1.6.0
numpy==2.1.2
opentelemetry-api==1.28.2
opentelemetry-sdk==1.28.2
opentelemetry-semantic-conventions==0.49b2
packaging==24.1
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.6
plotly==5.24.1
pluggy==1.5.0
postgrest==0.17.2
prometheus_client==0.21.0
prompt_toolkit==3.0.48
propcache==0.2.0
protobuf==5.28.2
psutil==6.0.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.2.0
pytest==8.3.3
pytest-timeout==2.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
realtime==2.0.6
referencing==0.35.1
requests==2.31.0
rich==13.9.2
rpds-py==0.20.0
rsa==4.7.2
s3transfer==0.10.3
scikit-learn==1.4.2
scipy==1.14.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
sqlparse==0.5.2
stack-data==0.6.3
starlette==0.40.0
storage3==0.8.2
streamlit==1.39.0
supabase==2.9.1
supafunc==0.6.2
tenacity==9.0.0
threadpoolctl==3.5.0
toml==0.10.2
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.31.1
wcwidth==0.2.13
websockets==13.1
Werkzeug==3.1.3
wrapt==1.17.0
yarl==1.15.5
zipp==3.21.0
